release: v0.34.2 — pipeline disk cache for Vulkan + DX12 (#331) (#346)
Extends existing in-memory shader cache to persistent disk storage.

Vulkan: VkPipelineCache per device, saved on destroy, restored on init.
DX12: GetCachedBlob/CachedPSO per-PSO blobs, keyed by SHA-256 of
root signature + shader bytecode + fixed-function state.

internal/pipelinecache: shared atomic blob I/O (write-tmp + rename).
Pipeline cache init is non-fatal — failure logs warning, continues uncached.

Also: lint fixes (errorlint, unconvert, hugeParam), race hardening
(sync.Once for RegisterHALBackends, mutex for adapter enumeration).

Contributor: @lkmavi

+++ hal/dx12/pso_cache_key.go
@@ -132,7 +132,7 @@ func writeGraphicsFixedState(
 	binary.LittleEndian.PutUint32(buf[0:4], uint32(psoDesc.PrimitiveTopologyType))
 	binary.LittleEndian.PutUint32(buf[4:8], psoDesc.NumRenderTargets)
 	binary.LittleEndian.PutUint32(buf[8:12], uint32(psoDesc.DSVFormat))
-	binary.LittleEndian.PutUint32(buf[12:16], uint32(psoDesc.SampleDesc.Count))
+	binary.LittleEndian.PutUint32(buf[12:16], psoDesc.SampleDesc.Count)
 	binary.LittleEndian.PutUint32(buf[16:20], psoDesc.SampleDesc.Quality)
 	binary.LittleEndian.PutUint32(buf[20:24], uint32(psoDesc.IBStripCutValue))
 	_, _ = h.Write(buf[:24])
